refactor: move old apis to memos v0
