+++ app/src/main/java/me/mudkip/moememos/data/api/MemosV1Api.kt
@@ -1,0 +1,5 @@
+package me.mudkip.moememos.data.api
+
+interface MemosV1Api {
+
+}
