+++ app/src/main/java/me/mudkip/moememos/data/datasource/ExplorePagingSource.kt
@@ -4,8 +4,8 @@ import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.skydoves.sandwich.getOrElse
 import com.skydoves.sandwich.mapSuccess
+import me.mudkip.moememos.data.api.MemosV0Memo
 import me.mudkip.moememos.data.constant.MoeMemosException
-import me.mudkip.moememos.data.model.Memo
 import me.mudkip.moememos.data.repository.MemoRepository
 import me.mudkip.moememos.ext.getErrorMessage
 
@@ -13,16 +13,16 @@ const val EXPLORE_PAGE_SIZE = 20
 
 class ExplorePagingSource(
     private val memoRepository: MemoRepository
-) : PagingSource<Int, Memo>() {
+) : PagingSource<Int, MemosV0Memo>() {
 
-    override fun getRefreshKey(state: PagingState<Int, Memo>): Int? {
+    override fun getRefreshKey(state: PagingState<Int, MemosV0Memo>): Int? {
         return state.anchorPosition?.let {
             state.closestPageToPosition(it)?.prevKey?.plus(1)
                 ?: state.closestPageToPosition(it)?.nextKey?.minus(1)
         }
     }
 
-    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Memo> {
+    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, MemosV0Memo> {
         val pageNumber = params.key ?: 0
         val response = memoRepository.listAllMemo(EXPLORE_PAGE_SIZE, pageNumber * EXPLORE_PAGE_SIZE)
