--- app/src/main/java/me/mudkip/moememos/data/model/Status.kt
@@ -1,14 +1,0 @@
-package me.mudkip.moememos.data.model
-
-import androidx.annotation.Keep
-
-@Keep
-data class Profile(
-    val mode: String,
-    val version: String
-)
-
-@Keep
-data class Status(
-    val profile: Profile
-)
