+++ app/src/main/java/me/mudkip/moememos/ui/component/Attachment.kt
@@ -13,13 +13,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.unit.dp
 import me.mudkip.moememos.R
-import me.mudkip.moememos.data.model.Resource
+import me.mudkip.moememos.data.api.MemosV0Resource
 import me.mudkip.moememos.ext.string
 import me.mudkip.moememos.viewmodel.LocalUserState
 
 @Composable
 fun Attachment(
-    resource: Resource
+    resource: MemosV0Resource
 ) {
     val uriHandler = LocalUriHandler.current
     val userStateViewModel = LocalUserState.current
