+++ app/src/main/java/me/mudkip/moememos/ui/component/MemosCard.kt
@@ -45,8 +45,8 @@ import androidx.compose.ui.unit.dp
 import com.skydoves.sandwich.suspendOnSuccess
 import kotlinx.coroutines.launch
 import me.mudkip.moememos.R
-import me.mudkip.moememos.data.model.Memo
-import me.mudkip.moememos.data.model.MemosRowStatus
+import me.mudkip.moememos.data.api.MemosRowStatus
+import me.mudkip.moememos.data.api.MemosV0Memo
 import me.mudkip.moememos.ext.icon
 import me.mudkip.moememos.ext.string
 import me.mudkip.moememos.ext.titleResource
@@ -57,7 +57,7 @@ import me.mudkip.moememos.viewmodel.LocalUserState
 
 @Composable
 fun MemosCard(
-    memo: Memo,
+    memo: MemosV0Memo,
     previewMode: Boolean = false
 ) {
     val memosViewModel = LocalMemos.current
@@ -109,7 +109,7 @@ fun MemosCard(
 
 @Composable
 fun MemosCardActionButton(
-    memo: Memo
+    memo: MemosV0Memo
 ) {
     var menuExpanded by remember { mutableStateOf(false) }
     val context = LocalContext.current
