+++ app/src/main/java/me/mudkip/moememos/viewmodel/ArchivedMemoListViewModel.kt
@@ -11,8 +11,8 @@ import com.skydoves.sandwich.suspendOnSuccess
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import me.mudkip.moememos.data.model.Memo
-import me.mudkip.moememos.data.model.MemosRowStatus
+import me.mudkip.moememos.data.api.MemosRowStatus
+import me.mudkip.moememos.data.api.MemosV0Memo
 import me.mudkip.moememos.data.repository.MemoRepository
 import me.mudkip.moememos.ext.string
 import me.mudkip.moememos.ext.suspendOnErrorMessage
@@ -22,7 +22,7 @@ import javax.inject.Inject
 class ArchivedMemoListViewModel @Inject constructor(
     private val memoRepository: MemoRepository
 ) : ViewModel() {
-    var memos = mutableStateListOf<Memo>()
+    var memos = mutableStateListOf<MemosV0Memo>()
         private set
 
     var errorMessage: String? by mutableStateOf(null)
