+++ app/src/main/java/me/mudkip/moememos/viewmodel/ResourceListViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.skydoves.sandwich.suspendOnSuccess
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import me.mudkip.moememos.data.model.Resource
+import me.mudkip.moememos.data.api.MemosV0Resource
 import me.mudkip.moememos.data.repository.ResourceRepository
 import javax.inject.Inject
 
@@ -14,7 +14,7 @@ import javax.inject.Inject
 class ResourceListViewModel @Inject constructor(
     private val resourceRepository: ResourceRepository
 ): ViewModel() {
-    var resources = mutableStateListOf<Resource>()
+    var resources = mutableStateListOf<MemosV0Resource>()
         private set
 
     fun loadResources() = viewModelScope.launch {
